backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Pillow==10.2.0
google-generativeai==0.3.2
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
PyJWT==2.8.0